implemented state change to display current tags or edit tags
