+++ src/components/posts/PostTags/CurrentPostTags.js
@@ -1,0 +1,11 @@
+import React from "react"
+
+export const CurrentPostTags = () => {
+
+    return (
+        <div className="current-post-tags">
+            <h4>#tag</h4>
+            <h4>#tag</h4>
+        </div>
+    )
+}
