+++ src/components/posts/PostTags/EditPostTags.js
@@ -1,0 +1,11 @@
+import React from "react"
+
+export const EditPostTags = () => {
+
+    return (
+        <div className="modify-tags-container">
+            <input type="checkbox" id="first-tag"></input>
+            <label for="#first-tag">#tag</label>
+        </div>
+    )
+}
